old_revs/rev4/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
pillow==11.0.0
PyQt6==6.7.1
PyQt6-Charts==6.7.0
PyQt6-Charts-Qt6==6.7.3
PyQt6-Qt6==6.7.3
PyQt6_sip==13.8.0
PySide6==6.8.0.2
PySide6_Addons==6.8.0.2
PySide6_Essentials==6.8.0.2
requests==2.32.3
shiboken6==6.8.0.2
urllib3==2.2.3
